24/11/2019

+++ electronics-sales/src/main/java/com/electronicssales/models/dtos/ProductDto.java
@@ -33,6 +33,6 @@ public class ProductDto {
 
     MultipartFile[] images;
 
-    Long quantity;
+    Integer quantity;
     
 }
